pyproject.toml
[project]
name = "cima-dot-com-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentql>=1.17.1",
    "fastapi>=0.128.0",
    "playwright>=1.57.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "uvicorn[standard]>=0.40.0",
]
